requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==22.2.0
importlib-metadata==4.8.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.1.1
pytest==7.0.1
pytest-html==3.2.0
pytest-metadata==1.11.0
tomli==1.2.3
typing_extensions==4.1.1
zipp==3.6.0
